♻️ 型代わりに利用する定数オブジェクトを別ファイルに移動
