+++ packages/rainbow-line/src/App.tsx
@@ -1,9 +1,15 @@
 import { useState } from "react";
 import { FiltersSample } from "./components/parts/filters";
 import { HeaderPlaceHolder } from "./components/parts/ph-header";
+import { FILTER_ATTRIBUTES } from "./interfaces/filter-attributes";
 
 function App() {
-  const [filters, setFilters] = useState({
+  const [filters, setFilters] = useState<
+    Record<
+      (typeof FILTER_ATTRIBUTES)[number]["id"],
+      (typeof FILTER_ATTRIBUTES)[number]["items"][number]["value"]
+    >
+  >({
     parkingLot: "all",
     region: "all",
     prefecture: "all",
@@ -18,6 +24,7 @@ function App() {
         defaultValues={filters}
         onFilterChange={(k, v) => setFilters({ ...filters, [`${k}`]: v })}
       />
+      <span>{JSON.stringify(filters)}</span>
     </div>
   );
 }
